% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_OneData.R
\name{BiomCorrMatrix}
\alias{BiomCorrMatrix}
\title{Generate the Correlation Matrix for the biomarkers in a dataset}
\usage{
BiomCorrMatrix(input, parameters)
}
\arguments{
\item{input}{A list of objects created with \code{MakeInput}}

\item{parameters}{A data.frame of objects created with \code{MakeParameters}}
}
\description{
A function that will create a correlation matrix using GenerateCorrM to be
used in the generation of a dataset
}
\examples{
input <- MakeInput(n_biom = 5, typeCorr = "AR1", rho = 0.4)
parameters <- MakeParameters(input=input)
BiomCorrMatrix(input, parameters)

}

