% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{ci.cox.lasso}
\alias{ci.cox.lasso}
\title{Obtains the PITE and confidence interval for an additional subject
using the Lasso output}
\usage{
ci.cox.lasso(dataset, alpha = 0.05, reg.results, reg.out, tol.beta = 0.1,
  gridrange_ = 100)
}
\arguments{
\item{dataset}{a data.frame}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}

\item{reg.results}{an object from the score.cox.lasso function}

\item{tol.beta}{Tolerance for determining if a coefficient is zero. to be passed to the fixedLassoInf}

\item{gridrange_}{Grid range for constructing confidence intervals, on the standardized scale. to be passed to the fixedLassoInf}
}
\description{
A function that takes the results from score.cox.lasso to get the PITE
and the confidence intervals for one subject.
}

