% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{ci.cox}
\alias{ci.cox}
\title{Obtains the PITE and confidence interval for an additional subject}
\usage{
ci.cox(dataset, reg.results, alpha = 0.05)
}
\arguments{
\item{dataset}{a data.frame}

\item{reg.results}{an object from the score.cox function}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}
}
\description{
A function that takes the results from score.cox to get the PITE
and the confidence intervals for one subject.
}

