% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{ci.cox.reduced}
\alias{ci.cox.reduced}
\title{Obtains the PITE and confidence interval for an additional subject
using the reduced model from the Lasso output}
\usage{
ci.cox.reduced(dataset, vars, reg.results, reg.out, n_biom = 6,
  alpha = 0.05)
}
\arguments{
\item{dataset}{a data.frame}

\item{vars}{List of selected variables.}

\item{reg.results}{an object resulting from score.cox.reduced}

\item{n_biom}{number of considered biomarkers}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}
}
\description{
A function that takes the results from score.cox.reduced to get the PITE
and the confidence intervals for one subject.
}

