% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Conf-Intervals.R
\name{confidence.intervals.lasso.test}
\alias{confidence.intervals.lasso.test}
\title{Confidence Intervals for Lasso}
\usage{
confidence.intervals.lasso.test(dataset, input, parameters, lasso.results,
  alpha = 0.05, gridrange_ = 250, tol.beta = 0.01)
}
\arguments{
\item{dataset}{A data.frame created with \code{OneData}}

\item{input}{A list of objects created with \code{MakeInput}}

\item{parameters}{A \code{MakeParameters} object}

\item{lasso.results}{The results obtained from score.lasso}

\item{alpha}{Significance level for confidence intervals (target is miscoverage alpha/2 in each tail)}

\item{gridrange_}{Grid range for constructing confidence intervals, on the standardized scale. to be passed to the fixedLassoInf}

\item{tol.beta}{Tolerance for determining if a coefficient is zero. to be passed to the fixedLassoInf}
}
\description{
Calculates the confidence intervals for the score given a Lasso estimation
}
\examples{
input <- MakeInput()
parameters <- MakeParameters(input = input)
dataset <-OneData(input,parameters)
score.lasso(dataset, input)

}

