% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Conf-Intervals.R
\name{confidence.intervals.ml.test}
\alias{confidence.intervals.ml.test}
\title{Confidence Intervals for the full model}
\usage{
confidence.intervals.ml.test(dataset, input, parameters, ML.results,
  alpha = 0.05)
}
\arguments{
\item{dataset}{A data.frame created with \code{OneData}}

\item{input}{A list of objects created with \code{MakeInput}}

\item{parameters}{A \code{MakeParameters} object}

\item{ML.results}{The results obtained from score.lm}

\item{alpha}{Significance level for confidence intervals (target is miscoverage alpha/2 in each tail)}
}
\description{
Calculates the confidence intervals for the score given the full model
}
\examples{
input <- MakeInput()
parameters <- MakeParameters(input = input)
dataset <-OneData(input,parameters)
score.lm(dataset, input)

}

