% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectiveInference_funs.fixedCox_eta.R
\name{fixedCoxLassoInf_eta}
\alias{fixedCoxLassoInf_eta}
\title{Perform selective inference for a specified contrast}
\usage{
fixedCoxLassoInf_eta(x, y, status, beta, lambda, alpha = 0.1,
  type = c("partial"), tol.beta = 1e-05, tol.kkt = 0.1,
  gridrange = c(-100, 100), bits = NULL, verbose = FALSE,
  this.call = NULL, contrast)
}
\description{
This function is a modified version of the selectiveInference::fixedCoxLassoInf.
This version allows an extra parameter 'contrast' which is a matrix with the
contrast to make inference for.
}

