% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_OneData.R
\name{GenerateCorrMatrix}
\alias{GenerateCorrMatrix}
\title{Generate Correlation Matrix}
\usage{
GenerateCorrMatrix(nvar, type.corr = "I", rho, blocks, nvar.blocks,
  rho.blocks)
}
\arguments{
\item{nvar}{Number of variables that will determine the matrix dimension
(\code{nvar} x \code{nvar})}

\item{type.corr}{Type of correlation matrix to be generated. Options:
\code{"AR1"} create a first-order autoregressive correlation matrix;
\code{"I"} creates an identity matrix; \code{"CS"} creates a Compound
simmetry matrix; \code{"block"} creates a matrix with blocks of equally
correlated variables; and \code{"step"} is used if the 3 cuartiles are used
to dichotomize a variable and form 3 different variables.}

\item{rho}{Correlation coefficient if the matrix is specified to be AR1 or
CS. must be in the range (-1, 1)}

\item{blocks}{Number of blocks in the matrix in case
type.corr=\code{"block"}}

\item{nvar.blocks}{Vector of length = \code{blocks} with the number of
variables in each block}

\item{rho.blocks}{Vector of length = \code{blocks} with the correlation
coefficients for each block}
}
\description{
A generic function that will create a correlation matrix depending on the
specified structure and the parameters
}
\examples{
GenerateCorrMatrix(nvar = 3, type.corr = "I")
GenerateCorrMatrix(nvar = 5, type.corr = "AR1", rho = 0.3)
GenerateCorrMatrix(nvar = 5, type.corr = "CS", rho = 0.3)
GenerateCorrMatrix(nvar = 5, type.corr = "block", blocks = 2,
                   nvar.blocks = c(3, 2), rho.blocks = c(0.5, 0.3))

}

