% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_OneData.R
\name{MakeInput}
\alias{MakeInput}
\title{Input parameters for simulations}
\usage{
MakeInput(y_type = 2, nT = 50, nC = 50, nsim = 100, n_biom = 1,
  n_biom_pred = 1, typeCorr = c("I", "CS", "AR1", "block", "manual"),
  rho = 0, blocks = 1, nvarblocks1_ = 2, nvarblocks2_ = NA,
  nvarblocks3_ = NA, rhoblocks1_ = 0, rhoblocks2_ = 0, rhoblocks3_ = 0,
  correlations = FALSE, a_ = 0, b_ = 0, sigma = 1)
}
\arguments{
\item{y_type}{type of variable for the outcome variable: 1=Binary; 2=Normal}

\item{nT}{Number of subjects in the treatment group}

\item{nC}{Number of subjects in the control group}

\item{nsim}{Number of simulations to be performed}

\item{n_biom}{Number of biomarkers in the dataset}

\item{n_biom_pred}{Number of prognostic/predictive biomarkers}

\item{typeCorr}{Type of correlation matrix for the biomarkers: one of "I"
for independent biomarkers, "CS" for compound symetry,
  "AR1" for autocorrelated, "block" for a block matrix, or "manual"
  which allows specifying each correlation (to be used in shiny)}

\item{rho}{Correlation coefficient if the matrix is specified to be AR1 or CS.
must be in the range (-1, 1)}

\item{blocks}{Number of blocks in the matrix in case type.corr="block".}

\item{nvarblocks1_}{If typeCorr = "block"; number of variables in block 1}

\item{nvarblocks2_}{If typeCorr = "block"; number of variables in block 2}

\item{nvarblocks3_}{If typeCorr = "block"; number of variables in block 3}

\item{rhoblocks1_}{If typeCorr = "block"; correlation coefficient for
variables in block 1}

\item{rhoblocks2_}{If typeCorr = "block"; correlation coefficient for
variables in block 2}

\item{rhoblocks3_}{If typeCorr = "block"; correlation coefficient for
variables in block 3}

\item{correlations}{TRUE/FALSE. Only used in shiny to display an input for
manual correlation matrix}

\item{a_}{A number indicating the baseline response in the patients}

\item{b_}{A number indicating the treatment effect in the patients}

\item{sigma}{Standard deviation for the response variable y}
}
\description{
A function that will create a list of input values to be used in other functions
of this package. Mainly to specify the parameters in the simulation.
}
\examples{
input <- MakeInput(y_type = 2, nT = 500,
                 nC = 500, n_biom = 30, n_biom_pred = 2)
input

}

