% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_OneData.R
\name{MakeInputTable}
\alias{MakeInputTable}
\title{Creates a nice table with the information of all the parameters.}
\usage{
MakeInputTable(input)
}
\arguments{
\item{input}{A list of objects created with \code{MakeInput}}
}
\description{
\code{MakeInputTable} returns a table with the information of all the
 parameters to create the dataset
}
\examples{
input <- MakeInput()
MakeInputTable(input)

}

