% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_OneData.R
\name{MakeParameters}
\alias{MakeParameters}
\title{Parameters for biomarkers}
\usage{
MakeParameters(prognosis = numeric(), predictive = numeric(),
  types = numeric(), prevalence = numeric(), means = numeric(),
  stddev = numeric(), input)
}
\arguments{
\item{prognosis}{A vector of length \code{n_biom_pred} indicating the
prognostic coefficients in the model for the generation of the data}

\item{predictive}{A vector of length \code{n_biom_pred} indicating the
predictive coefficients in the model for the generation of the data}

\item{types}{A vector of length \code{n_biom_pred} indicating the type of
biomarker to be generated 1:Binary; 2:Normal, 3:Uniform. For uniform biomarkers
we use U(-sqrt(3), sqrt(3)) to have std.dev=1}

\item{prevalence}{A vector of length \code{n_biom_pred} indicating the
prevalence for the binary biomarkers. If the biomarker in the position i is
not binary, the ith element of this vector is ignored}

\item{means}{A vector of length \code{n_biom_pred} indicating the mean for
the NORMAL biomarkers. If the biomarker in the position i is not normal,
the ith element of this vector is ignored}

\item{stddev}{A vector of length \code{n_biom_pred} indicating the standard
deviation for the NORMAL biomarkers. If the biomarker in the position i is
not normal, the ith element of this vector is ignored}

\item{input}{an object created with \code{MakeInput} function}
}
\description{
A function that will create a data.frame with the parameters to be used in
the simulations. It is used to specify distributions and effects of biomarkers
}
\examples{
input <- MakeInput(y_type = 2, nT = 500, nC = 500,
                   n_biom = 30, n_biom_pred = 2)
MakeParameters(prognosis = c(1, 0), predictive = c(0, 1), types = c(1, 1),
  prevalence = c(0.5, 0), means = c(0, 0), stddev = c(1, 1), input = input)

}

