% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_OneData.R
\name{MakeParametersTable}
\alias{MakeParametersTable}
\title{Creates a nice table with the information of all the parameters}
\usage{
MakeParametersTable(input, parameters)
}
\arguments{
\item{input}{A list of objects created with \code{MakeInput}}

\item{parameters}{A list of objects created with \code{MakeParameters}}
}
\description{
\code{MakeParametersTable} returns a table with the information of all the
 biomarkers created in the dataset
}
\examples{
input<-MakeInput()
parameters<-MakeParameters(input = input)
MakeParametersTable(input,parameters)

}

