% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_OneData.R
\name{OneSubject}
\alias{OneSubject}
\title{Creates one simulated subject}
\usage{
OneSubject(input, parameters, standardize = FALSE, param = "REF")
}
\arguments{
\item{input}{A list of objects created with \code{MakeInput}}

\item{parameters}{A data.frame of objects created with \code{MakeParameters}}

\item{standardize}{logical. If TRUE all covariates are scaled to have 0 mean
and sd 1}

\item{param}{parametrization for the binary variables. Either "REF" or "EFFECT"}
}
\description{
A function that will create a data.frame with one subject with simulated
covariates
}
\examples{
input <- MakeInput()
parameters <- MakeParameters(input = input)

}

