% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{score.cox.lasso}
\alias{score.cox.lasso}
\title{Fits the full model (all treatment-covariate interactions) with lasso penalty
for model selection}
\usage{
score.cox.lasso(dataset, alpha = 0.05, calculate.ci = TRUE,
  lam_frac = 0.5, tol.beta = 0.1, gridrange_ = 100)
}
\arguments{
\item{dataset}{a data.frame}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}

\item{calculate.ci}{logical. whether to calculate the confidence intervals for pite}

\item{lam_frac}{multiplier for the lagrange parameter}

\item{tol.beta}{Tolerance for determining if a coefficient is zero. to be passed to the fixedLassoInf}

\item{gridrange_}{Grid range for constructing confidence intervals, on the standardized scale. to be passed to the fixedLassoInf}
}
\description{
A function that fits the model with treatment effect, prognostic terms
and interaction terms with glmnet
}

