% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{score.cox.null}
\alias{score.cox.null}
\title{Fits the 'null' model (ATE with no interactions)}
\usage{
score.cox.null(dataset, alpha = 0.05, calculate.ci = TRUE)
}
\arguments{
\item{dataset}{a data.frame}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}

\item{calculate.ci}{logical. whether to calculate the confidence intervals for pite}
}
\description{
A function that fits the model with treatment effect and prognostic terms
but no interaction terms
}

