% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{score.cox}
\alias{score.cox}
\title{Fits the full model (all treatment-covariate interactions)}
\usage{
score.cox(dataset, alpha = 0.05, calculate.ci = TRUE)
}
\arguments{
\item{dataset}{a data.frame}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}

\item{calculate.ci}{logical. whether to calculate the confidence intervals for pite}
}
\description{
A function that fits the model with treatment effect, prognostic terms
and interaction terms
}

