% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{score.cox.reduced}
\alias{score.cox.reduced}
\title{Fits the reduced model obtained with the lasso but with no penalization on
the paraemeters}
\usage{
score.cox.reduced(dataset, vars, n_biom = 6, alpha = 0.05,
  calculate.ci = TRUE)
}
\arguments{
\item{dataset}{a data.frame}

\item{vars}{List of selected variables.}

\item{n_biom}{number of considered biomarkers}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}

\item{calculate.ci}{logical. whether to calculate the confidence intervals for pite}
}
\description{
A function that fits the model with treatment effect, prognostic terms
and interaction terms with coxph using the model obtained by the lasso
}

