% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_randomized_lasso.R
\name{score.lasso.added.noise}
\alias{score.lasso.added.noise}
\title{Fit linear model to create a score using randomized Lasso}
\usage{
score.lasso.added.noise(dataset, input, parameters = NULL, verbose = TRUE,
  alpha = 0.05, lambda = "lagrange", lam_frac = 1, nfolds = 10,
  perturb_frac = 0.2, pite.ci = TRUE, n.pite = NULL, ndraw = 15000,
  burnin = 2000)
}
\arguments{
\item{dataset}{A data.frame created with \code{OneData}}

\item{input}{A list of objects created with \code{MakeInput}}

\item{parameters}{A \code{MakeParameters} object}

\item{verbose}{logical. whether to print a message with results.}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}

\item{lambda}{penalization parameter. one of "lagrange", "lambda.min", "lambda.1se" or a number}

\item{lam_frac}{multiplier for the lagrange parameter}

\item{nfolds}{number of folds for crossvalidations. to be passed to glmnet}

\item{perturb_frac}{scale of Gaussian noise added to the response. this number is then multiplied to sd(y)}

\item{pite.ci}{logical. whether to calculate ci for pite}

\item{n.pite}{a number specifying the number of subjects in the dataset for which to calculate the ci for the pite if pite.ci is TRUE}

\item{ndraw}{Number of samples of optimization variables to sample.}

\item{burnin}{How many samples of optimization variable to discard.}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}
}
\description{
Performs randomized lasso in a dataset with main effects and interactions.
calculates PITE and CI
}
\examples{
input <- MakeInput()
parameters <- MakeParameters(input = input)
dataset <- OneData(input, parameters)
score.lasso.added.noise(dataset, input, parameters)

}

