% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores.R
\name{score.lasso}
\alias{score.lasso}
\title{Fit linear model to create a score using LASSO}
\usage{
score.lasso(dataset, input, parameters = NULL, verbose = TRUE,
  alpha = 0.05, lam_frac = 1, lambda = "lambda.min", nfolds = 10,
  pite.ci = TRUE, gridrange_ = 100, tol.beta = 0.01)
}
\arguments{
\item{dataset}{A data.frame created with \code{OneData}}

\item{input}{A list of objects created with \code{MakeInput}}

\item{parameters}{A \code{MakeParameters} object}

\item{verbose}{logical. whether to print a message with results.}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}

\item{lam_frac}{multiplier for the lagrange parameter}

\item{lambda}{penalization parameter. one of "lagrange", "lambda.min", "lambda.1se" or a number}

\item{nfolds}{number of folds for crossvalidations. to be passed to glmnet}

\item{pite.ci}{logical. whether to calculate the ci for the pite.}

\item{gridrange_}{Grid range for constructing confidence intervals, on the standardized scale. to be passed to the fixedLassoInf}

\item{tol.beta}{Tolerance for determining if a coefficient is zero. to be passed to the fixedLassoInf}
}
\description{
Performs lasso in a dataset with main effects and interactions. calculates
PITE and CI
}
\examples{
input <- MakeInput()
parameters <- MakeParameters(input = input)
dataset <- OneData(input,parameters)
score.lasso(dataset, input)

}

