% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores.R
\name{score.lm}
\alias{score.lm}
\title{Fit linear model to create a score using glm}
\usage{
score.lm(dataset, input, verbose = TRUE, alpha = 0.05)
}
\arguments{
\item{dataset}{A data.frame created with \code{OneData}}

\item{input}{A list of objects created with \code{MakeInput}}

\item{verbose}{logical. whether to print a message with results.}

\item{alpha}{Significance level for confidence intervals (target is
miscoverage alpha/2 in each tail)}
}
\description{
Performs glm in a dataset with main effects and interactions. Calculates
PITE and CI
}
\examples{
input <- MakeInput()
parameters <- MakeParameters(input = input)
dataset <- OneData(input,parameters)
score.lm(dataset, input)

}

