% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_simulate-trial.R
\name{simulateNtrial_randomized}
\alias{simulateNtrial_randomized}
\title{Simulate datasets and Create score variables in each of them}
\usage{
simulateNtrial_randomized(nsim, effects, case, npergroup, mc.cores = 1,
  verbose = TRUE, n_biom = 2, param = "EFFECT", lambda = "lambda.min",
  nfolds = 10, typeCorr = "I", rho = 0, alpha = 0.05, test = FALSE,
  perturb_frac = c(0, 0.1, 0.2, 0.8, 1, 1.5, 2), par.method = "mclapply",
  pite.ci = TRUE, n.pite = NULL, n.test = NULL, lam_frac = 1)
}
\arguments{
\item{nsim}{number of simulated datasets to be generated}

\item{effects}{a 6-columns matrix of data.frame
with the parameters for a, b, g1, g2, d1 and d2 in the model.
The rows correspond to different cases to analyze}

\item{case}{Specify the row of 'effects' that needs to be evaluated}

\item{npergroup}{number of subjects per treatment arm}

\item{mc.cores}{number of cores for parallel computing. To be passed to mclapply}

\item{verbose}{logical. whether to display intermediate results.}

\item{n_biom}{number of biomarkers in the model}

\item{param}{Coding to be used. Only "EFFECT" is allowed now. Treatment and binary variables are coded -1/1}

\item{lambda}{penalization to be used for the lasso. Either "lambda.min", "lambda.1se", "lagrange", or a number}

\item{nfolds}{number of folds in crossvalidation for glmnet}

\item{typeCorr}{Correlation between biomarkers. See \link{input}. For "block" use simulateNtrial_block.}

\item{rho}{correlation between biomarkers when typeCorr is "AR1" or "CS"}

\item{alpha}{significance level to build the confidence intervals}

\item{test}{logical. whether to evaluate the score in a test dataset}

\item{perturb_frac}{a numeric vector of length 7 that indicates the degree of noise to introduce in the randomized lasso problem.}

\item{par.method}{Only "mclapply" is allowed}

\item{pite.ci}{logical. whether to calculate the ci for pite in the train data.}

\item{n.pite}{number of subjects for which to analyze coverage in the train data if pite.ci = TRUE}

\item{n.test}{numeric. The size of the test dataset.}

\item{lam_frac}{numeric. a fraction to multiply the lagrange parameter in the fixed lambda case.}

\item{tol.beta}{tolerance to determine a parameter estimate is 0. To be passed to the selectiveInference functions.}
}
\description{
Creates \code{nsim} datasets with specified parameters,
performs lm, lasso and randomized lasso in this dataset and creates PITE and its CI.
This function allows comparing randomized lasso with different degrees of noise.
It takes perturb_frac a vector of length 7 and performs seven model fits with the randomized
lasso.
Up to two biomarkers may have an effect on the response (mkr1 and mkr2).
}
\details{
This function performs the simulations for the specified cases.
}
\examples{
input <- MakeInput()
parameters <- MakeParameters(input = input)
dataset <- OneData(input,parameters)
score.lm(dataset, input)

}

