% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Summarize-Trial.R
\name{summarize_scores}
\alias{summarize_scores}
\title{Summarize the results from the scoring funcitons.}
\usage{
summarize_scores(scores, dataset, n.pite = 40)
}
\arguments{
\item{scores}{The dataset with the PITE and their confidence intervals}

\item{dataset}{The dataset with the clinical trial data}

\item{n.pite}{number of subjects to consider to summarize}
}
\description{
Calculates the width of the confidence intervals for pite, their coverage,
and the bias and mse of the point estimates.
Also calculates the sensitivity and specifity of subgroups.
}
\examples{
input <- MakeInput()
parameters <- MakeParameters(input = input)
dataset <-OneData(input,parameters)
score.lm(dataset, input)

}

