% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{SurvData}
\alias{SurvData}
\title{Creates a list with nsim datasets using the SurvOneData function}
\usage{
SurvData(alpha, beta, delta1, delta2, nsim, nsub, sigma = 1, maxtime = 2)
}
\arguments{
\item{alpha}{True value for intercept}

\item{beta}{True value for treatment effect}

\item{delta1}{True value for predictive effect of biomarker 1}

\item{delta2}{True value for predictive effect of biomarker 2}

\item{nsub}{number of subjects in dataset}

\item{sigma}{standard deviation for the biomarkers}

\item{maxtime}{maximum study time for censoring observations}
}
\description{
A function that will create list of data.frames with a simulated data
}

