% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{SurvOneData}
\alias{SurvOneData}
\title{Creates one simulated dataset with survival outcome}
\usage{
SurvOneData(alpha, beta, delta1, delta2, nsub, sigma = 0.2, maxtime = 2)
}
\arguments{
\item{alpha}{True value for intercept}

\item{beta}{True value for treatment effect}

\item{delta1}{True value for predictive effect of biomarker 1}

\item{delta2}{True value for predictive effect of biomarker 2}

\item{nsub}{number of subjects in dataset}

\item{sigma}{standard deviation for the biomarkers}

\item{maxtime}{maximum study time for censoring observations}
}
\description{
A function that will create a data.frame with simulated data
}

