% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_scores_survival.R
\name{SurvOneSubject}
\alias{SurvOneSubject}
\title{Creates one simulated dataset with survival outcome with only one subject
for test in the simulations}
\usage{
SurvOneSubject(alpha, beta, delta1, delta2, sigma = 0.2, maxtime = 2)
}
\arguments{
\item{alpha}{True value for intercept}

\item{beta}{True value for treatment effect}

\item{delta1}{True value for predictive effect of biomarker 1}

\item{delta2}{True value for predictive effect of biomarker 2}

\item{sigma}{standard deviation for the biomarkers}

\item{maxtime}{maximum study time for censoring observations}
}
\description{
A function that will create a data.frame with simulated data for one subject
}

